% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{read_cai}
\alias{read_cai}
\title{Parse a file from EMBOSS cai program.}
\usage{
read_cai(path)
}
\arguments{
\item{path}{A string containing the path to a cai file.}
}
\value{
A dataframe containing the parsed data with column headings (seqid, cai).
}
\description{
Parse a file from EMBOSS cai program.
}
\examples{
# Load some sample data included with the package.
cai_path <- system.file("extdata", "lmac.cai", package = "codonfriend")
# Parse that file.
cai_table <- read_cai(cai_path)
head(cai_table)

}
