% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{read_gff3}
\alias{read_gff3}
\title{Read a GFF3 file as a GRanges object..}
\usage{
read_gff3(path)
}
\arguments{
\item{path}{The path to the GFF file.}
}
\value{
A GenomicRanges `GRanges` object.
}
\description{
Read a GFF3 file as a GRanges object..
}
\examples{

path <- system.file("extdata", "Lepmu1.gff3", package = "codonfriend")
gff <- read_gff3(path)
head(gff)

}
