% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{read_many}
\alias{read_many}
\title{Read many files and combine them.}
\usage{
read_many(paths, FUN = read_cai, colname = "file")
}
\arguments{
\item{paths}{A vector of files to parse.}

\item{FUN}{The function to use to parse the files. Default `read_cai`.}

\item{colname}{What to call the column added.}
}
\value{
A concatenated dataframe with filenames added as a new column.
}
\description{
Read many files and combine them.
}
\examples{

paths <- list.files("dir", full.names = TRUE)
# or
paths <- Sys.glob("dir/*.cai")
df <- read_many(paths, FUN=read_cai, colname="file")
head(df)

}
